+++ package.json
@@ -25,7 +25,7 @@
     "react-native-gesture-handler": "~2.9.0",
     "react-native-image-picker": "^5.6.0",
     "react-native-reanimated": "~2.14.4",
-    "react-native-safe-area-context": "4.5.0",
+    "react-native-safe-area-context": "^4.5.0",
     "react-native-screens": "~3.20.0",
     "react-native-smooth-pincode-input": "^1.0.9",
     "react-redux": "^8.1.1"
